Fix compile from merge
