+++ include/pokemon.h
@@ -568,7 +568,6 @@ extern const struct SpriteTemplate gBattlerSpriteTemplates[];
 extern const u32 sExpCandyExperienceTable[];
 extern const struct AbilityInfo gAbilitiesInfo[];
 extern const struct NatureInfo gNaturesInfo[];
-extern const struct BehaviorOWE gOWESpeciesBehavior[OWE_SPECIES_BEHAVIOR_COUNT];
 
 static inline const u8 *GetAbilityName(enum Ability ability)
 {
@@ -769,54 +768,5 @@ struct BoxPokemon *GetSelectedBoxMonFromPcOrParty(void);
 u32 GiveScriptedMonToPlayer(struct Pokemon *mon, u8 slot);
 void ChangePokemonNicknameWithCallback(void (*callback)(void));
 bool32 HasShedinjaHPHandling(enum Species species);
-
-static inline u32 OWE_GetMovementTypeFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].movementType;
-}
-
-static inline u32 OWE_GetViewDistanceFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].viewDistance;
-}
-
-static inline u32 OWE_GetViewWidthFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].viewWidth;
-}
-
-static inline u32 OWE_GetViewActiveDistanceFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].activeDistance;
-}
-
-static inline enum SpeedOWE OWE_GetIdleSpeedFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].idleSpeed;
-}
-
-static inline enum SpeedOWE OWE_GetActiveSpeedFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].activeSpeed;
-}
-
-static inline enum ReturnToIdleOWE OWE_GetReturnToIdleFromSpecies(enum Species speciesId)
-{
-    speciesId = SanitizeSpeciesId(speciesId);
-    enum OverworldWildEncounterBehaviors behavior = gSpeciesInfo[speciesId].overworldEncounterBehavior;
-    return gOWESpeciesBehavior[behavior].returnToIdle;
-}
 
 #endif // GUARD_POKEMON_H
